requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.3.1
environs==9.5.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
packaging==23.1
python-dotenv==1.0.0
pytz==2023.3
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
django==4.2.4
psycopg2-binary~=2.9.3
asgiref~=3.7.2
redis==4.6.0
ujson==5.8.0